refactor: 당첨 번호에 관한 기능 클래스 분리
