+++ src/test/java/lotto/temp/LottoTempTest.java
@@ -19,16 +19,6 @@ class LottoTempTest {
     }
 
     @Test
-    @DisplayName("입력받은 당첨번호 콤마를 기준으로 분리")
-    void separateWinningNumber(){
-        assertThat(lottoTemp.createWinningNumbers("1,2,3,4,5,6"))
-                .containsExactly(1,2,3,4,5,6);
-        assertThat(lottoTemp.createWinningNumbers("3,5,14,26,34,45"))
-                .containsExactly(3,5,14,26,34,45);
-    }
-
-
-    @Test
     @DisplayName("구입한 로또와 당첨번호가 몇 개 일치하는지 확인하는 테스트")
     void compareLottos(){
         assertThat(lottoTemp.compareLottos(
